+++ palace/drivers/basesolver.hpp
@@ -17,7 +17,7 @@ class FiniteElementSpaceHierarchy;
 class IoData;
 class Mesh;
 class Timer;
-class PortExcitationHelper;
+class PortExcitations;
 
 //
 // Base driver class for all simulation types.
@@ -52,7 +52,7 @@ public:
   template <typename SolverType>
   void SaveMetadata(const SolverType &ksp) const;
   void SaveMetadata(const Timer &timer) const;
-  void SaveMetadata(const PortExcitationHelper &excitation_helper) const;
+  void SaveMetadata(const PortExcitations &excitation_helper) const;
 };
 
 }  // namespace palace