+++ palace/models/portexcitations.hpp
@@ -7,14 +7,16 @@
 #include <map>
 #include <vector>
 #include <nlohmann/json_fwd.hpp>
-#include "utils/strongtype.hpp"
 
 namespace palace
 {
+class LumpedPortOperator;
+class WavePortOperator;
+class SurfaceCurrentOperator;
 
 // Small helper class to collect data of what (lumped / wave / surface) ports are
 // excited in driven and transient simulation, as stored in space_op;
-// Manges indices.
+// Manages indices.
 
 enum class PortType : std::uint8_t
 {
@@ -24,7 +26,7 @@ enum class PortType : std::uint8_t
   Undefined = 3
 };
 
-class PortExcitationHelper
+class PortExcitations
 {
 public:
   struct SingleExcitationSpec
@@ -34,7 +36,7 @@ public:
     std::vector<int> current_port = {};
 
     // TODO: C++20 to replace this with iterator over joined range.
-    auto flatten_port_indices() const
+    auto FlatternPortIndices() const
     {
       std::vector<int> out;
       out.insert(out.end(), lumped_port.cbegin(), lumped_port.cend());
@@ -44,7 +46,7 @@ public:
     }
 
     // Only a single port is excited.
-    std::tuple<bool, PortType, int> is_simple() const
+    std::tuple<bool, PortType, int> IsSimple() const
     {
       auto n_lumped = lumped_port.size();
       auto n_wave = wave_port.size();
@@ -69,62 +71,21 @@ public:
     }
   };
 
-  std::map<ExcitationIdx, SingleExcitationSpec> excitations = {};
+  std::map<int, SingleExcitationSpec> excitations = {};
 
   auto begin() { return excitations.begin(); }
   auto end() { return excitations.end(); }
   auto begin() const { return excitations.begin(); }
   auto end() const { return excitations.end(); }
 
-  template <typename LumpedportIt, typename WaveportIt, typename CurrentIt>
-  PortExcitationHelper(const LumpedportIt &lumped_port_op, const WaveportIt &wave_port_op,
-                       const CurrentIt &surf_j_op)
-  {
-    for (const auto &[idx, port] : lumped_port_op)
-    {
-      if (port.excitation ==
-          ExcitationIdx(0))  // LumpedPortData does not have a HasExcitation() function
-      {
-        continue;
-      }
-      excitations.try_emplace(port.excitation, SingleExcitationSpec{});  // If not present
-      excitations.at(port.excitation).lumped_port.push_back(idx);
-    }
-    for (const auto &[idx, port] : wave_port_op)
-    {
-      if (port.excitation ==
-          ExcitationIdx(0))  // WavePortData does not have a HasExcitation() function
-      {
-        continue;
-      }
-      excitations.try_emplace(port.excitation, SingleExcitationSpec{});
-      excitations.at(port.excitation).wave_port.push_back(idx);
-    }
+  PortExcitations(const LumpedPortOperator &lumped_port_op,
+                  const WavePortOperator &wave_port_op,
+                  const SurfaceCurrentOperator &surf_j_op);
 
-    // Surface currents are always excited. Add them to all single existing excitations.
-    // TODO: Add excitation 1 if not present already?
-    std::vector<int> current_port_idx;
-    for (const auto &[idx, port] : surf_j_op)
-    {
-      current_port_idx.push_back(idx);
-    }
-    if (!current_port_idx.empty())
-    {
-      for (auto &[ex_idx, ex_spec] : excitations)
-      {
-        ex_spec.current_port = current_port_idx;
-      }
-    }
-  };
-
-  [[nodiscard]] ExcitationIdx MaxIdx() const
+  [[nodiscard]] int MaxIdx() const
   {
-    if (excitations.empty())
-    {
-      return ExcitationIdx(0);
-    }
     // Map is stored order by key so max key is last item.
-    return std::next(std::rend(excitations))->first;
+    return excitations.empty() ? 0 : std::next(std::rend(excitations))->first;
   }
   [[nodiscard]] auto Size() const { return excitations.size(); }
   [[nodiscard]] auto Empty() const { return excitations.empty(); }
@@ -132,35 +93,35 @@ public:
   [[nodiscard]] std::string FmtLog() const;
 
   // Single Simple (only 1 port per excitation) Excitation.
-  [[nodiscard]] std::tuple<bool, ExcitationIdx, PortType, int> IsSingleSimple() const
+  [[nodiscard]] std::tuple<bool, int, PortType, int> IsSingleSimple() const
   {
     if (Size() == 1)
     {
       const auto &[ex_idx, ex_spec] = *excitations.begin();
-      const auto [is_simple, port_type, port_idx] = ex_spec.is_simple();
+      const auto [is_simple, port_type, port_idx] = ex_spec.IsSimple();
       if (is_simple)
       {
         return std::make_tuple(true, ex_idx, port_type, port_idx);
       }
     }
-    return std::make_tuple(false, ExcitationIdx(0), PortType::Undefined, 0);
+    return std::make_tuple(false, 0, PortType::Undefined, 0);
   }
 
   // Multiple Simple (only 1 port per excitation) Excitation.
   [[nodiscard]] bool IsMultipleSimple() const
   {
     return std::all_of(excitations.begin(), excitations.end(),
-                       [](const auto &ex) { return std::get<0>(ex.second.is_simple()); });
+                       [](const auto &ex) { return std::get<0>(ex.second.IsSimple()); });
   }
 };
 
-void to_json(nlohmann::json &j, const PortExcitationHelper::SingleExcitationSpec &p);
+void to_json(nlohmann::json &j, const PortExcitations::SingleExcitationSpec &p);
 
-void from_json(const nlohmann::json &j, PortExcitationHelper::SingleExcitationSpec &p);
+void from_json(const nlohmann::json &j, PortExcitations::SingleExcitationSpec &p);
 
-void to_json(nlohmann::json &j, const PortExcitationHelper &p);
+void to_json(nlohmann::json &j, const PortExcitations &p);
 
-void from_json(const nlohmann::json &j, PortExcitationHelper &p);
+void from_json(const nlohmann::json &j, PortExcitations &p);
 
 }  // namespace palace