+++ palace/models/timeoperator.cpp
@@ -8,7 +8,7 @@
 #include "linalg/iterative.hpp"
 #include "linalg/jacobi.hpp"
 #include "linalg/solver.hpp"
-#include "models/portexcitationhelper.hpp"
+#include "models/portexcitations.hpp"
 #include "models/spaceoperator.hpp"
 #include "utils/communication.hpp"
 #include "utils/iodata.hpp"
@@ -66,7 +66,7 @@ public:
     M = space_op.GetMassMatrix<Operator>(Operator::DIAG_ONE);
 
     // Already asserted that only that time dependant solver only has a single excitation.
-    auto excitation_helper = space_op.GetPortExcitationHelper();
+    auto excitation_helper = space_op.GetPortExcitations();
     auto excitation_idx = excitation_helper.excitations.begin()->first;
     // Set up RHS vector for the current source term: -g'(t) J, where g(t) handles the time
     // dependence.
@@ -286,7 +286,7 @@ TimeOperator::TimeOperator(const IoData &iodata, SpaceOperator &space_op,
   : rel_tol(iodata.solver.transient.rel_tol), abs_tol(iodata.solver.transient.abs_tol),
     order(iodata.solver.transient.order)
 {
-  auto excitation_helper = space_op.GetPortExcitationHelper();
+  auto excitation_helper = space_op.GetPortExcitations();
   // Should have already asserted that time dependant solver only has a single excitation.
   MFEM_VERIFY(excitation_helper.Size() == 1,
               fmt::format("Transient evolution currently only allows for a single "