+++ CHANGELOG.md
@@ -75,8 +75,8 @@ The format of this changelog is based on
   - Changed boundary indexing to enforce that the value of `Index` is unique between `LumpedPort`,
     `WavePort`, `SurfaceCurrent` and `Terminal`. This is a breaking change — e.g. currently a lumped
     port and a wave port could share an index.
-  - Added mutli-excitation support. `Excitation` in `LumpedPort` and `WavePort` boundary conditions
-    can not be specified either as bools or excitation integers, but not both at the same time.
+  - Added multi-excitation support. `Excitation` in `LumpedPort` and `WavePort` boundary conditions
+    can now be specified either as bools or excitation integers, but not both at the same time.
     Allow driven simulations to have multiple excitations indices in same configuration file to be
     simulated consecutively during the same Palace run. The measurement output of multiple
     excitations is printed to the same csv files with distinguish columns being post-indexed by the
@@ -86,7 +86,7 @@ The format of this changelog is based on
     excitation, but the PROM is shared between excitations.
   - Added general strong-type class `StrongT` and use it for excitation index `ExcitationIdx` to
     ensure it is not accidentally converted to port index.
-  - Added `PortExcitaitonHelper` to manage excitation pattern and print it to json metadata.
+  - Added `PortExcitationHelper` to manage excitation pattern and print it to json metadata.
 
 ## [0.13.0] - 2024-05-20